Revert "Merge branch 'main' into bichlien"
This reverts commit 43f7339bf3d2f686fe26657f74810a22819958c2, reversing
changes made to ca9e543ea5b913695fc57fae3ff7b30df6317d8a.

+++ src/main/java/com/example/eSmartRecruit/repositories/BlacklistRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.Blacklist;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.Blacklist;
-
 @Repository
-public interface BlacklistRepos extends JpaRepository<Blacklist, Long>{
+public interface BlacklistRepos extends JpaRepository<Blacklist, Integer>{
 
 }
