+++ src/main/java/com/example/eSmartRecruit/repositories/CommunicationRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.Communication;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.Communication;
-
 @Repository
-public interface CommunicationRepos extends JpaRepository<Communication, Long>{
+public interface CommunicationRepos extends JpaRepository<Communication, Integer>{
 
 }
