+++ src/main/java/com/example/eSmartRecruit/repositories/PositionRepos.java
@@ -1,11 +1,10 @@
 package com.example.eSmartRecruit.repositories;
 
+import com.example.eSmartRecruit.models.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.eSmartRecruit.models.Position;
-
 @Repository
-public interface PositionRepos extends JpaRepository<Position, Long>{
+public interface PositionRepos extends JpaRepository<Position, Integer>{
 
 }
